requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
pandas 
numpy
matplotlib
seaborn
orjson
urllib3>=1.26.0
scikit-learn>=1.0.0 
requests>=2.31.0
python-dotenv
tensorflow
tensorflow-hub
fastapi
uvicorn
streamlit
groq


